+++ adm/app/adms/Views/sitsUser/viewSitsUser.php
@@ -11,7 +11,7 @@
     echo "<a href='".URLADM."list-sits-users/index'>Listar</a><br>";
     if(!empty($this->data['viewSitUser'])){
         echo "<a href='" . URLADM . "edit-sits-users/index/" . $this->data['viewSitUser'][0]['id'] . "'>Editar</a><br>";
-        echo "<a href='" . URLADM . "delete-sits-users/index/" . $this->data['viewSitUser'][0]['id'] . "'>Apagar</a><br><br>";
+        echo "<a href='" . URLADM . "delete-sits-users/index/" . $this->data['viewSitUser'][0]['id'] . "' onclick='return confirm(\"Tem certeza que deseja apagar este ítem?\")>Apagar</a><br><br>";
     }
 
     if(isset($_SESSION['msg'])){
